+++ PixiEditor/Helpers/Converters/ViewboxInverseTransformConverter.cs
@@ -15,7 +15,7 @@ namespace PixiEditor.Helpers.Converters
             return transform.Inverse;
         }
 
-        public object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
+        public override object[] ConvertBack(object value, Type[] targetTypes, object parameter, CultureInfo culture)
         {
             throw new NotImplementedException();
         }
