+++ PixiEditor/Helpers/Extensions/PixiParserHelper.cs
@@ -5,7 +5,7 @@ namespace PixiEditor.Helpers.Extensions
 {
     public static class PixiParserHelper
     {
-        public static SKRectI GetRect(this SerializableLayer layer) => 
+        public static SKRectI GetRect(this SerializableLayer layer) =>
             SKRectI.Create(layer.OffsetX, layer.OffsetY, layer.Width, layer.Height);
     }
 }
